requirements.txt
dm-meltingpot==2.2.0
dmlab2d==1.0.0
gym==0.22.0
torch==2.2.2
tensorflow==2.11.1
pygame==2.5.2
protobuf==3.19.6
numpy==1.24.3
jax==0.4.12
jaxlib==0.4.12
gymnasium==0.26.3
chex==0.1.7
dm-env==1.6
dm-tree==0.1.8
immutabledict==4.2.0
ml-collections==0.1.1
networkx==3.3
opencv-python==4.9.0.80
pandas==2.2.2
pygame==2.5.2
reactivex==4.0.4
tensorflow-probability==0.24.0
ray[all]==2.6.1
matplotlib==3.8.4
pydantic==1.10.12
wandb==0.16.6